SOX-35 Rename PermissionCheckerInterface -> PermissionCheckInterface
